backend/requirements.txt
# requirements.txt
fastapi
uvicorn
python-dotenv
openai
firebase-admin
python-multipart
